correct issue minerva an nav from the tablet
Former-commit-id: 1f8f0739e8687d74ae0a3c17a1d007a8eff60ec2

+++ src/apps/xplorer/App.h
@@ -273,9 +273,9 @@ private:
     ///Light model for the scene
     osg::ref_ptr< osg::LightModel > light_model_0;
     ///-90 z up matrix
-    gmtl::Matrix44f mZUp;
+    gmtl::Matrix44d mZUp;
     ///User nav position for camera
-    gmtl::Matrix44f mNavPosition;
+    gmtl::Matrix44d mNavPosition;
     
     ///
     ves::xplorer::SceneRenderToTexturePtr mSceneRenderToTexture;
